[refactor] move group-related settings source to subdirectory
