+++ src/settings/SettingsView.ts
@@ -15,7 +15,7 @@ import {UserListView} from "./UserListView"
 import type {ReceivedGroupInvitation, User} from "../api/entities/sys/TypeRefs.js"
 import {CustomerInfoTypeRef, UserTypeRef} from "../api/entities/sys/TypeRefs.js"
 import {logins} from "../api/main/LoginController"
-import {GroupListView} from "./GroupListView"
+import {GroupListView} from "./groups/GroupListView.js"
 import {ContactFormListView} from "./contactform/ContactFormListView.js"
 import {WhitelabelSettingsViewer} from "./whitelabel/WhitelabelSettingsViewer"
 import {Icons} from "../gui/base/icons/Icons"
