+++ src/settings/UserViewer.ts
@@ -12,7 +12,7 @@ import {logins} from "../api/main/LoginController"
 import type {ContactForm} from "../api/entities/tutanota/TypeRefs.js"
 import {ContactFormTypeRef, CustomerContactFormGroupRootTypeRef, MailboxGroupRootTypeRef} from "../api/entities/tutanota/TypeRefs.js"
 import {ColumnWidth, Table, TableAttrs} from "../gui/base/Table.js"
-import {getGroupTypeName} from "./GroupViewer"
+import {getGroupTypeName} from "./groups/GroupViewer.js"
 import {Icons} from "../gui/base/icons/Icons"
 import {SecondFactorsEditForm} from "./login/secondfactor/SecondFactorsEditForm.js"
 import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
