+++ src/settings/groups/AddGroupDialog.ts
@@ -1,24 +1,24 @@
 import m, {Children, Component, Vnode} from "mithril"
-import {BookingItemFeatureType, FeatureType, GroupType} from "../api/common/TutanotaConstants"
-import {Dialog} from "../gui/base/Dialog"
-import type {ValidationResult} from "./SelectMailAddressForm"
-import {SelectMailAddressForm} from "./SelectMailAddressForm"
-import {getGroupTypeName} from "./GroupViewer"
-import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
-import {logins} from "../api/main/LoginController"
-import type {TranslationKey} from "../misc/LanguageViewModel"
-import {lang} from "../misc/LanguageViewModel"
-import {showBuyDialog} from "../subscription/BuyDialog"
-import {PreconditionFailedError} from "../api/common/error/RestError"
-import {showBusinessFeatureRequiredDialog} from "../misc/SubscriptionDialogs"
-import {TemplateGroupPreconditionFailedReason} from "../sharing/GroupUtils"
-import {DropDownSelector} from "../gui/base/DropDownSelector.js"
-import {TextField} from "../gui/base/TextField.js"
+import {BookingItemFeatureType, FeatureType, GroupType} from "../../api/common/TutanotaConstants.js"
+import {Dialog} from "../../gui/base/Dialog.js"
+import type {ValidationResult} from "../SelectMailAddressForm.js"
+import {SelectMailAddressForm} from "../SelectMailAddressForm.js"
+import {getGroupTypeName} from "./GroupViewer.js"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog.js"
+import {logins} from "../../api/main/LoginController.js"
+import type {TranslationKey} from "../../misc/LanguageViewModel.js"
+import {lang} from "../../misc/LanguageViewModel.js"
+import {showBuyDialog} from "../../subscription/BuyDialog.js"
+import {PreconditionFailedError} from "../../api/common/error/RestError.js"
+import {showBusinessFeatureRequiredDialog} from "../../misc/SubscriptionDialogs.js"
+import {TemplateGroupPreconditionFailedReason} from "../../sharing/GroupUtils.js"
+import {DropDownSelector} from "../../gui/base/DropDownSelector.js"
+import {TextField} from "../../gui/base/TextField.js"
 import {firstThrow, ofClass} from "@tutao/tutanota-utils"
-import type {GroupManagementFacade} from "../api/worker/facades/GroupManagementFacade"
-import {locator} from "../api/main/MainLocator"
-import {assertMainOrNode} from "../api/common/Env"
-import {getAvailableDomains} from "./mailaddress/MailAddressesUtils.js"
+import type {GroupManagementFacade} from "../../api/worker/facades/GroupManagementFacade.js"
+import {locator} from "../../api/main/MainLocator.js"
+import {assertMainOrNode} from "../../api/common/Env.js"
+import {getAvailableDomains} from "../mailaddress/MailAddressesUtils.js"
 
 assertMainOrNode()
