+++ src/settings/groups/GroupListView.ts
@@ -1,36 +1,36 @@
 import m, {Children} from "mithril"
-import type {VirtualRow} from "../gui/base/List"
-import {List} from "../gui/base/List"
-import {lang} from "../misc/LanguageViewModel"
-import {NotFoundError} from "../api/common/error/RestError"
-import {size} from "../gui/size"
-import type {GroupInfo} from "../api/entities/sys/TypeRefs.js"
-import {GroupInfoTypeRef} from "../api/entities/sys/TypeRefs.js"
-import {CustomerTypeRef} from "../api/entities/sys/TypeRefs.js"
+import type {VirtualRow} from "../../gui/base/List.js"
+import {List} from "../../gui/base/List.js"
+import {lang} from "../../misc/LanguageViewModel.js"
+import {NotFoundError} from "../../api/common/error/RestError.js"
+import {size} from "../../gui/size.js"
+import type {GroupInfo} from "../../api/entities/sys/TypeRefs.js"
+import {GroupInfoTypeRef} from "../../api/entities/sys/TypeRefs.js"
+import {CustomerTypeRef} from "../../api/entities/sys/TypeRefs.js"
 import {neverNull, noOp} from "@tutao/tutanota-utils"
-import type {SettingsView, UpdatableSettingsViewer} from "./SettingsView"
+import type {SettingsView, UpdatableSettingsViewer} from "../SettingsView.js"
 import {LazyLoaded} from "@tutao/tutanota-utils"
-import {logins} from "../api/main/LoginController"
-import {GroupViewer} from "./GroupViewer"
-import * as AddGroupDialog from "./AddGroupDialog"
-import {Icon} from "../gui/base/Icon"
-import {Icons} from "../gui/base/icons/Icons"
-import {OperationType} from "../api/common/TutanotaConstants"
-import {BootIcons} from "../gui/base/icons/BootIcons"
-import {header} from "../gui/Header.js"
-import {isAdministratedGroup} from "../search/model/SearchUtils"
-import {GroupMemberTypeRef} from "../api/entities/sys/TypeRefs.js"
-import type {EntityUpdateData} from "../api/main/EventController"
-import {isUpdateForTypeRef} from "../api/main/EventController"
-import {Button, ButtonType} from "../gui/base/Button.js"
-import type {GroupMembership} from "../api/entities/sys/TypeRefs.js"
-import {compareGroupInfos} from "../api/common/utils/GroupUtils"
-import {GENERATED_MAX_ID} from "../api/common/utils/EntityUtils"
-import {showNotAvailableForFreeDialog} from "../misc/SubscriptionDialogs"
-import {locator} from "../api/main/MainLocator"
-import {ListColumnWrapper} from "../gui/ListColumnWrapper"
+import {logins} from "../../api/main/LoginController.js"
+import {GroupViewer} from "./GroupViewer.js"
+import * as AddGroupDialog from "./AddGroupDialog.js"
+import {Icon} from "../../gui/base/Icon.js"
+import {Icons} from "../../gui/base/icons/Icons.js"
+import {OperationType} from "../../api/common/TutanotaConstants.js"
+import {BootIcons} from "../../gui/base/icons/BootIcons.js"
+import {header} from "../../gui/Header.js"
+import {isAdministratedGroup} from "../../search/model/SearchUtils.js"
+import {GroupMemberTypeRef} from "../../api/entities/sys/TypeRefs.js"
+import type {EntityUpdateData} from "../../api/main/EventController.js"
+import {isUpdateForTypeRef} from "../../api/main/EventController.js"
+import {Button, ButtonType} from "../../gui/base/Button.js"
+import type {GroupMembership} from "../../api/entities/sys/TypeRefs.js"
+import {compareGroupInfos} from "../../api/common/utils/GroupUtils.js"
+import {GENERATED_MAX_ID} from "../../api/common/utils/EntityUtils.js"
+import {showNotAvailableForFreeDialog} from "../../misc/SubscriptionDialogs.js"
+import {locator} from "../../api/main/MainLocator.js"
+import {ListColumnWrapper} from "../../gui/ListColumnWrapper.js"
 import {ofClass, promiseMap} from "@tutao/tutanota-utils"
-import {assertMainOrNode} from "../api/common/Env"
+import {assertMainOrNode} from "../../api/common/Env.js"
 import Stream from "mithril/stream";
 
 assertMainOrNode()
