+++ src/settings/groups/GroupViewer.ts
@@ -1,33 +1,33 @@
 import m, {Children} from "mithril"
-import {Dialog} from "../gui/base/Dialog"
-import {formatDateWithMonth, formatStorageSize} from "../misc/Formatter"
-import {lang} from "../misc/LanguageViewModel"
+import {Dialog} from "../../gui/base/Dialog.js"
+import {formatDateWithMonth, formatStorageSize} from "../../misc/Formatter.js"
+import {lang} from "../../misc/LanguageViewModel.js"
 import {assertNotNull, firstThrow, LazyLoaded, neverNull, ofClass, promiseMap} from "@tutao/tutanota-utils"
-import type {Group, GroupInfo} from "../api/entities/sys/TypeRefs.js"
-import {AdministratedGroupTypeRef, CustomerTypeRef, GroupInfoTypeRef, GroupMemberTypeRef, GroupTypeRef, UserTypeRef} from "../api/entities/sys/TypeRefs.js"
-import {BookingItemFeatureType, GroupType, OperationType} from "../api/common/TutanotaConstants"
-import {BadRequestError, NotAuthorizedError, PreconditionFailedError} from "../api/common/error/RestError"
-import type {TableAttrs} from "../gui/base/Table.js"
-import {ColumnWidth, Table, TableLineAttrs} from "../gui/base/Table.js"
-import {logins} from "../api/main/LoginController"
-import {Icons} from "../gui/base/icons/Icons"
-import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
-import {localAdminGroupInfoModel} from "./LocalAdminGroupInfoModel"
-import type {EntityUpdateData} from "../api/main/EventController"
-import {isUpdateForTypeRef} from "../api/main/EventController"
-import {compareGroupInfos, getGroupInfoDisplayName} from "../api/common/utils/GroupUtils"
-import {GENERATED_MAX_ID, GENERATED_MIN_ID, isSameId} from "../api/common/utils/EntityUtils"
-import {showBuyDialog} from "../subscription/BuyDialog"
-import type {TextFieldAttrs} from "../gui/base/TextField.js"
-import {TextField} from "../gui/base/TextField.js"
-import type {DropDownSelectorAttrs, SelectorItemList} from "../gui/base/DropDownSelector.js"
-import {DropDownSelector} from "../gui/base/DropDownSelector.js"
-import type {EntityClient} from "../api/common/EntityClient"
-import type {UpdatableSettingsDetailsViewer} from "./SettingsView"
-import {locator} from "../api/main/MainLocator"
-import {assertMainOrNode} from "../api/common/Env"
-import {IconButton, IconButtonAttrs} from "../gui/base/IconButton.js"
-import {ButtonSize} from "../gui/base/ButtonSize.js"
+import type {Group, GroupInfo} from "../../api/entities/sys/TypeRefs.js"
+import {AdministratedGroupTypeRef, CustomerTypeRef, GroupInfoTypeRef, GroupMemberTypeRef, GroupTypeRef, UserTypeRef} from "../../api/entities/sys/TypeRefs.js"
+import {BookingItemFeatureType, GroupType, OperationType} from "../../api/common/TutanotaConstants.js"
+import {BadRequestError, NotAuthorizedError, PreconditionFailedError} from "../../api/common/error/RestError.js"
+import type {TableAttrs} from "../../gui/base/Table.js"
+import {ColumnWidth, Table, TableLineAttrs} from "../../gui/base/Table.js"
+import {logins} from "../../api/main/LoginController.js"
+import {Icons} from "../../gui/base/icons/Icons.js"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog.js"
+import {localAdminGroupInfoModel} from "../LocalAdminGroupInfoModel.js"
+import type {EntityUpdateData} from "../../api/main/EventController.js"
+import {isUpdateForTypeRef} from "../../api/main/EventController.js"
+import {compareGroupInfos, getGroupInfoDisplayName} from "../../api/common/utils/GroupUtils.js"
+import {GENERATED_MAX_ID, GENERATED_MIN_ID, isSameId} from "../../api/common/utils/EntityUtils.js"
+import {showBuyDialog} from "../../subscription/BuyDialog.js"
+import type {TextFieldAttrs} from "../../gui/base/TextField.js"
+import {TextField} from "../../gui/base/TextField.js"
+import type {DropDownSelectorAttrs, SelectorItemList} from "../../gui/base/DropDownSelector.js"
+import {DropDownSelector} from "../../gui/base/DropDownSelector.js"
+import type {EntityClient} from "../../api/common/EntityClient.js"
+import type {UpdatableSettingsDetailsViewer} from "../SettingsView.js"
+import {locator} from "../../api/main/MainLocator.js"
+import {assertMainOrNode} from "../../api/common/Env.js"
+import {IconButton, IconButtonAttrs} from "../../gui/base/IconButton.js"
+import {ButtonSize} from "../../gui/base/ButtonSize.js"
 
 assertMainOrNode()
